booking_app/requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.7.2
billiard==4.2.1
bootstrap4==0.1.0
build==1.2.2.post1
CacheControl==0.14.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crashtest==0.4.1
distlib==0.3.9
Django==4.2.16
django-extensions==3.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
kombu==5.5.4
more-itertools==10.7.0
msgpack==1.1.0
packaging==25.0
pbs-installer==2025.4.9
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.2
poetry-core==2.1.2
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
RapidFuzz==3.13.0
redis==6.2.0
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.2
trove-classifiers==2025.4.11.15
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13
xattr==1.1.4
zstandard==0.23.0
psycopg2-binary==2.9.10
django-environ==0.12.0

